requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.0
pydantic==2.5.2
motor==3.3.2